requirements.txt
streamlit
openai
langchain
langchain-openai
langchain-community
faiss-cpu
numpy
pymupdf
pdfplumber
python-dotenv
